Completes ex. 13.4, makes app web/HTTP-based
This commit transforms the application into a web-based application
relying upon express. It supports GET to /api/blogs, which responds
with the list of blogs in the SQL database; POST to /api/blogs, which
adds the request body to the database if formatted correctly, or
responds with status code 400 and the error message if not; and
DELETE to /api/blogs/:id, which destroys the entry at 'id', or if it
does not exist responds with status code 404.