Completes ex. 13.5, refactors app into components
This commit refactors the application into multiple component parts.
The file models/blog.js contains the SQL model for a blog entry, and
the file models/index.js syncs all models with the database. The file
util/db.js handles the connection to the database, and the file
util/config.js provides the configuration information for the
connection. Finally, controllers/blogs.js contains all of the API
functions for interacting with the database.