Completes ex. 13.8, adds users to app & DB
This commit adds a bloglistusers collection to the database via the
application, with connection to the blogs collection so that each
blog entry has a bloglistuser_id value. A bloglistuser entry has an
key id, and required strings for username (must be unique), name, and
a password hash. The bloglistuser entries also have created_at and
updated_at timestamps.

The commit adds POST and GET routes at /api/users, the first of which
takes username, name, and password data in the request body and adds
a new user to the collection; and the second of which displays all
users. It also adds a PUT route at /api/users/:username which takes a
new username in the request body and modifies the username
appropriately.