Completes ex 13.16, adds /api/authors route
This commit adds a GET route to /api/authors which sends an SQL query
to the database that groups by author, and provides the total number
of articles by the author and the sum total of likes for those
articles. It also sorts the response in descending order of total
likes.