Completes ex 13.18, adds Year field via migration
This commit adds a migration to add a year field to the blog model
with a custom validator function to ensure the year is between 1991
and present.

This also commit also makes updates to the blogs controller to
account for the new year field, and also to require a user to be
logged in to add a new blog. It also makes updats to allow for
rolling back migrations.

Finally, the original migration is updated due to some typos that
caused problems.