Completes es 13.19&20, adds readling list table
This commit adds a mutation which creates a readling_lists table to
the DB. This table is a connection table between bloglistusers and
blogs, which also includes a boolean member 'read' which is whether
the blog has been read yet. The commit adds a PUSH route to
/api/readinglists to add readinglist entries.

The commit also enables a GET route at /api/users/:id which displays
a user's username, name, and reading list.