Completes ex. 13.24, adds server-side sessions
This commit implements server-side session token management for the
bloglist application. Via migrations, it adds a new bloglist_sessions
collection to the database, each member of which has a token string
and the user id of the associated user. The login route now adds a
session to the collection upon successful login.

Logout functionality is implemented as a DELETE route to /api/logout,
requiring a header with the username of the logging out user. The
logout function finds all tokens stored in the database collection
matching that user and deletes them (accomplished via a scope
function in the bloglist_sessions collection definition).

Token validation is moved into middleware, and also now validates
that the token matches one present in the database.